Run update_feed to update channel_list
Bit of a headache I ran into while trying to get this working: On my thinkpad, nothing in the Documents (or Pictures, Videos, Music, Favorites) folder can be edited with a script thanks to "Ransomware Protection". I cannot edit the list of protected folders, nor can I allow a program to pass through the protection.